+++ stubs/DecodeLabs/Archetype.php
@@ -34,9 +34,13 @@ class Archetype implements Proxy
     public static function extend(string $interface, string $namespace): void
     {
     }
-    public static function resolve(string $interface, string $name, callable|string|null $default = null): string
+    public static function resolve(string $interface, ?string $name = null, callable|string|null $default = null): string
     {
         return static::$instance->resolve(...func_get_args());
+    }
+    public static function tryResolve(string $interface, ?string $name = null, callable|string|null $default = null): ?string
+    {
+        return static::$instance->tryResolve(...func_get_args());
     }
     public static function normalize(string $interface, string $name): string
     {